Remove references to ikvm in services names
The ikvm-viewer service is renamed for viewer-app in order to support
more than ikvm without having to create new services or groups.  This
would be the name of this service if there was a base image.

The next step would be to extract everything non related to ikvm in
another project and only keep here what is specific to the Supermicro
IPMI console.